+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All changes and updates, that are relevant for a user will be documented here
 ### Bug Fix
 
 - Account was not found anymore when creating a new one and then saving. Only after restarting of the app it was again visible.
+- Remove unknown currency selection when creating a new account
 
 ### Features